seek in local file

+++ openplayer/openplayer/InputStreamConnection.h
@@ -26,6 +26,6 @@
 -(id)initWithUrl:(NSURL *)url;
 -(long)readData:(uint8_t *)buffer maxLength:(NSUInteger) length;
 -(void)closeStream;
--(BOOL)skip:(long)offset;
+-(BOOL)skip:(float)offset;
 
 @end
